pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/everything_in_its_place",
]

[tool.hatch.build.targets.wheel]
packages = [
    "src/everything_in_its_place",
]

[project]
name = "a-place-for-everything-and-everything-in-its-place"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name="Geoffrey Woollard", email="geoffwoollard@gmail.com" },
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "omegaconf",
  "hydra-core",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[project.urls]
Documentation = "https://github.com/geoffwoollard/everything_in_its_place/blob/main/README.md"

[tool.hatch.version]
path = "src/everything_in_its_place/__about__.py"

[tool.coverage.run]
source_pkgs = ["everything_in_its_place", "tests"]
branch = true
parallel = true


[tool.coverage.paths]
everything_in_its_place = ["src/everything_in_its_place", "*/everything_in_its_place/src/everything_in_its_place"]
tests = ["tests", "*/everything_in_its_place/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[project.scripts]
everything_in_its_place = "everything_in_its_place.run:main"